% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makes2btable.R
\name{makes2btable}
\alias{makes2btable}
\title{Organize table of results for S2B function}
\usage{
makes2btable(complete_results, apid_main, A_index, B_index)
}
\arguments{
\item{complete_results}{data frame with S2B and specificity scores}

\item{apid_main}{igraph object with the network analyzed by S2B}

\item{A_index}{vector with seed ids}

\item{B_index}{vector with seed ids}
}
\value{
data frame with S2B results and node annotation
}
\description{
Function called by S2B
}
