% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multineibstat2.R
\name{multineibstat2}
\alias{multineibstat2}
\title{identifies nodes with direct neigbohrs in seeds lists}
\usage{
multineibstat2(g, A_index, B_index, nrep)
}
\arguments{
\item{g}{igraph object with the network analyzed by S2B}

\item{A_index}{vector with seed ids}

\item{B_index}{vector with seed ids}

\item{nrep}{number of network randomizations (shufle egdes maintaining degree) used to compute specificity score}
}
\value{
data frame with number of direct neighbors in both seed lists, the related number of crossbridges and its specificity score
}
\description{
Function called by makes2btable to annotate S2B results
}
