% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bthreshold.R
\name{s2bthreshold}
\alias{s2bthreshold}
\title{S2B threshold}
\usage{
s2bthreshold(s2bvec)
}
\arguments{
\item{s2bvec}{vector of S2B scores obtained with S2B or subS2B.}
}
\value{
s2bt - numeric value of the S2B threshold
}
\description{
This function defines the S2B threshold that separates nodes with high/low S2B score
}
