% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedrows.R
\name{seedrows}
\alias{seedrows}
\title{Find set of node ids in graph}
\usage{
seedrows(seed_graph, seedvec)
}
\arguments{
\item{seed_graph}{igraph object of the network to be searched.}

\item{seedvec}{vector with strings of node identifiers.}
}
\value{
rowindex vector of integers with the found node ids
}
\description{
This function allows you to get a vector of node ids matching your input vector with strings of node identifiers. Identifiers not found in the network are ignored.
}
