% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpmain.R
\name{simpmain}
\alias{simpmain}
\title{Simplify network}
\usage{
simpmain(seed_graph)
}
\arguments{
\item{seed_graph}{igraph object of the network to be simplified.}
}
\value{
seed_graph_maincomp is an igraph object with the simplified main component of the input graph.
}
\description{
This function allows you to remove multiple edges, self-loops and keeps only main component.
}
