% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subS2Btrack.R
\name{subS2Btrack}
\alias{subS2Btrack}
\title{S2B auxiliary function}
\usage{
subS2Btrack(seed_graph, index1, index2, meandist)
}
\arguments{
\item{seed_graph}{igraph object of the network to be searched.}

\item{index1}{vector with integer node identifiers.}

\item{index2}{vector with integer node identifiers.}

\item{meandist}{length threshold above which shortest paths are not included in S2B computation}
}
\value{
allcount - vector with S2B scores for each node in the graph

smat1 - matrix with one row for each node in the network and a column for each seed in index1. If the matrix element (i,j) is 1, then index1[j] was connected to the other seed set through a shortest path going through node i, and shorter than meandist

smat2 - matrix similar to smat1 but reffering to index 2 seeds

maxS2B - theoretical maximum of absolute S2B scores for this network and seed sets

trackpath - list composed of two vectors with the ids of the shortest path extremes and a list with the shortest path members
}
\description{
This function computes S2B scores, without specificity scores, and keeps track of wich shortest paths were used during the computation.
}
